Created starting page
I have created the first page, note that the picture of the logo is random and serves only positioning purposes. The colors and styles can be later changed if it doesn't fit our discription. Further more it does fit the wireframe description. The buttons do not have any backend yet nore are they functional just yet. If you guys want to do some more I encourage you to  create folders of "Models", "Controllers" and "Views" so we have everything in order.